Fix menus in "watch" mode
In sculpin's watch mode (`vendor/bin/sculpin generate --watch --server`), the navigation menu seems to get blanked out on pages that are being edited. From debugging the process in flight, it seems that this is because the `->updatedSources()` method will only return the 1 file being changed, and this doesn't result in a valid menu being assembled.

By switching to `->allSources()`, the bug is fixed, albeit at the performance cost of reapplying the menu for each format run.